✨ Create dependabot.yml
This commit introduces a dependabot.yml configuration file to the project. Dependabot is a tool that helps us keep our dependencies up-to-date. It periodically checks if there are updates available for our dependencies and automatically opens pull requests if it finds any. This is crucial for maintaining the security and stability of our project, as outdated dependencies can introduce vulnerabilities and compatibility issues.

By adding this file, we’re instructing Dependabot on how to perform these checks. This includes information about the package managers to use, the frequency of updates, and the directories to check. With this commit, we’re taking a significant step towards improving the maintainability of our project.